admin-ui: Fix AnswerCategoryCounts initial map value wrong type bug
Initial map value should be int64, instead of int.

+++ module/admin-api/ELKStats/services/visitStatsService.go
@@ -649,9 +649,9 @@ func AnswerCategoryCounts(ctx context.Context, client *elastic.Client,
 	}
 
 	answerCategoryCounts := make(map[string]interface{})
-	answerCategoryCounts[businessFilterName] = 0
-	answerCategoryCounts[chatFilterName] = 0
-	answerCategoryCounts[otherFilterName] = 0
+	answerCategoryCounts[businessFilterName] = int64(0)
+	answerCategoryCounts[chatFilterName] = int64(0)
+	answerCategoryCounts[otherFilterName] = int64(0)
 
 	if agg, found := result.Aggregations.Filters(aggName); found {
 		businessBucket, ok := agg.NamedBuckets[businessFilterName]
